make auto save an HOC
