+++ src/components/SongRouter.tsx
@@ -3,13 +3,15 @@ import { Route, useHistory } from "react-router-dom";
 import { ChordSong } from "../common/ChordModel/ChordSong";
 import Song, { SongMode } from "./Song";
 
-interface SongProps {
+interface SongRouterProps {
     basePath: string;
     initialSong: ChordSong;
     onSongChanged?: (song: ChordSong) => void;
 }
 
-const SongRouter: React.FC<SongProps> = (props: SongProps): JSX.Element => {
+const SongRouter: React.FC<SongRouterProps> = (
+    props: SongRouterProps
+): JSX.Element => {
     const history = useHistory();
 
     const editPath = `${props.basePath}/edit`;
@@ -28,7 +30,6 @@ const SongRouter: React.FC<SongProps> = (props: SongProps): JSX.Element => {
             <Route key={editPath} path={editPath}>
                 <Song
                     initialSong={props.initialSong}
-                    basePath={props.basePath}
                     mode={SongMode.Edit}
                     onSongChanged={props.onSongChanged}
                     onEdit={switchToEdit}
@@ -38,7 +39,6 @@ const SongRouter: React.FC<SongProps> = (props: SongProps): JSX.Element => {
             <Route key={playPath} path={playPath}>
                 <Song
                     initialSong={props.initialSong}
-                    basePath={props.basePath}
                     mode={SongMode.Play}
                     onSongChanged={props.onSongChanged}
                     onEdit={switchToEdit}
